requirements.txt
pyyaml==5.1.2
aiohttp==3.6.2
pubnub==4.1.6
pycryptodome==3.9.0
